pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "anima-slam-coko"
version = "0.1.0"
description = "ANIMA SLAM-COKO — Compact Keyframe-Optimized Multi-Agent Gaussian Splatting SLAM"
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy>=1.26",
    "opencv-python>=4.10",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
]

[project.optional-dependencies]
dev = [
    "mypy>=1.11",
    "pytest>=8.3",
    "ruff>=0.11",
]
mlx = ["mlx>=0.22; platform_system == 'Darwin'"]
paper = [
    "open3d>=0.19",
    "pillow>=10.4",
    "pyyaml>=6.0",
    "rich>=13.9",
    "transformers>=4.46",
    "typer>=0.15",
]
cuda = [
    "torch>=2.5",
    "torchaudio>=2.5",
    "torchvision>=0.20",
]

[tool.hatch.build.targets.wheel]
packages = ["src/anima_slam_coko"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
